pyproject.toml
[tool.poetry]
name = "airflow_dags"
version = "0.1.0"
description = ""
authors = ["smezzacapo <stephenmezzo@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
apache-airflow = "^1.10.10"

[tool.poetry.dev-dependencies]
pytest = "^5.4.3"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
